+++ DateBar/DateBarWidget.cpp
@@ -1,6 +1,5 @@
 // SPDX-FileCopyrightText: 2003-2020 The KPhotoAlbum Development Team
-// SPDX-FileCopyrightText: 2021 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
-// SPDX-FileCopyrightText: 2022 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2021-2023 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -378,18 +377,19 @@ void DateBar::DateBarWidget::drawHistograms(QPainter &p)
     }
 
     int unit = 0;
-    const int minUnit = unitForDate(m_dates->lowerLimit());
-    const int maxUnit = (unitForDate(m_dates->upperLimit()) != -1) ? unitForDate(m_dates->upperLimit()) : numberOfUnits();
+    const int minUnit = unitForDate(m_dates->lowerLimit()); // first non-empty unit
+    const int maxUnit = (unitForDate(m_dates->upperLimit()) != -1) ? unitForDate(m_dates->upperLimit()) : numberOfUnits(); // last non-empty unit
     const bool linearScale = Settings::SettingsData::instance()->histogramUseLinearScale();
     for (int x = rect.x(); x + m_barWidth < rect.right(); x += m_barWidth, unit += 1) {
-        if (unit <= minUnit || unit > maxUnit) {
+        if (unit < minUnit || unit > maxUnit) {
             Qt::BrushStyle style = Qt::SolidPattern;
 
             p.setBrush(QBrush(Qt::lightGray, style));
             p.drawRect(x, 1, m_barWidth, rect.height() + 2);
             continue;
         }
-        const DB::ImageCount count = m_dates->count(rangeForUnit(unit));
+        const auto unitRange = rangeForUnit(unit);
+        const DB::ImageCount count = m_dates->count(unitRange);
         int exactPx = 0;
         int rangePx = 0;
         if (max != 0) {